enhance msp and network creation function to make it able to reuse from external as fabric cli
